% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicPermTest.R
\name{classicPermTest}
\alias{classicPermTest}
\title{Classic Permutation Test for RNA-Seq Data}
\usage{
classicPermTest(dge, design, nPerm)
}
\arguments{
\item{dge}{A dgeList object, created with edgeR, containing the normalization factors as computed by edgeR.}

\item{design}{A model matrix; The first column should be all 1s. The second column should have two unique values,
corresponding to the groups}

\item{nPerm}{Number of random permutations used for the computation of the p-value}
}
\description{
Computes p-values using a classic permutation test based on the absolute
difference in means.
}
\examples{
Y <- rnbinom(20*10, mu = 10, size = 1/0.2)
Y <- data.frame(array(Y, dim = c(20, 10)))
X1 <- as.factor(rep(c("A", "B"), each = 20/2))
design <- model.matrix(~X1, contrasts.arg = list(X1 = "contr.sum"))
dge <- edgeR::DGEList(counts = t(Y), group = X1)
dge <- edgeR::calcNormFactors(dge)
pClassic <- classicPermTest(dge, design, 200)
}
\references{
Hemerik, Jesse, and Jelle Goeman. "Exact testing with random permutations." Test 27.4 (2018): 811-825.
}
\author{
Jakob Walter
}
