% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipScoresTest.R
\name{FlipScoresTest}
\alias{FlipScoresTest}
\title{Classic Permutation Test for RNA-Seq Data}
\usage{
FlipScoresTest(dge, design, scoreType = "basic", toBeTested = 2, nPerm = 5000)
}
\arguments{
\item{nPerm}{Number of random permutations used for the computation of the p-value}

\item{X}{A vector encoded as a factor with two levels.
The levels encode the two different classes that are to be tested for differential expression.}

\item{Y}{An array or dataframe}
}
\description{
Computes p-values using a classic permutation test based on the absolute
difference in means. The data is normalized using edgeR.
}
\examples{
Y <- rnbinom(40*100, mu = 10, size = 1/0.2)
Y <- data.frame(array(Y, dim = c(40, 100)))
X <- as.factor(rep(c("A", "B"), each = 20))
FlipScoresTestNoCovBasic (X, Y, 100)
}
\keyword{Internal}
