% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWUTest.R
\name{MWUTest}
\alias{MWUTest}
\title{MWU-Test for RNA-Seq Data}
\usage{
MWUTest(dge, design)
}
\arguments{
\item{X}{A vector encoded as a factor with two levels.
The levels encode the two different classes that are to be tested for differential expression.}

\item{Y}{An array or dataframe}
}
\description{
Computes p-values using a Mann Whitney U-test. The data is normalized using edgeR.
}
\examples{
Y <- rnbinom(40*100, mu = 10, size = 1/0.2)
Y <- data.frame(array(Y, dim = c(40, 100)))
X <- as.factor(rep(c("A", "B"), each = 20))
MWUTest(X, Y, 100)
}
