% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateNB.R
\name{SimulateNB}
\alias{SimulateNB}
\title{Simulating RNA-Seq Data from NB distribution}
\usage{
SimulateNB(
  nSamples,
  nGenes,
  offset = rep(1, nSamples),
  mu1,
  mu2,
  phi1,
  phi2,
  beta = rep(0, nGenes)
)
}
\arguments{
\item{nSamples}{number of generated samples. Must be divisible by two as our
group sizes are equal}

\item{nGenes}{number of generated genes.}
}
\description{
This functions allows you to generate RNA-Seq like data from a negative
binomial distribution.
}
\examples{
cat_function()
}
\keyword{cats}
