% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicPermTest.R
\name{classicPermTest}
\alias{classicPermTest}
\title{Classic Permutation Test for RNA-Seq Data}
\usage{
classicPermTest(dge, design, nPerm)
}
\arguments{
\item{nPerm}{Number of random permutations used for the computation of the p-value}

\item{X}{A vector encoded as a factor with two levels.
The levels encode the two different classes that are to be tested for differential expression.}

\item{Y}{An array or dataframe}
}
\description{
Computes p-values using a classic permutation test based on the absolute
difference in means. The data is normalized using edgeR.
}
\examples{
Y <- rnbinom(40*100, mu = 10, size = 1/0.2)
Y <- data.frame(array(Y, dim = c(40, 100)))
X <- as.factor(rep(c("A", "B"), each = 20))
ClassicPermTest(X, Y, 100)
}
