% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWUTest.R
\name{MWUTest}
\alias{MWUTest}
\title{MWU test for RNA-Seq Data}
\usage{
MWUTest(dge, design)
}
\arguments{
\item{dge}{A dgeList object, created with edgeR, containing the normalization factors as computed by edgeR.}

\item{design}{A model matrix; The first column should be all 1s. The second column should have two unique values,
corresponding to the groups}
}
\description{
Computes p-values using a Mann-Whitney U Test for the null hypothesis \eqn{H_0: \beta_j = 0}.
}
\examples{
Y <- rnbinom(20*10, mu = 10, size = 1/0.2)
Y <- data.frame(array(Y, dim = c(20, 10)))
X1 <- as.factor(rep(c("A", "B"), each = 20/2))
design <- model.matrix(~X1, contrasts.arg = list(X1 = "contr.sum"))
dge <- edgeR::DGEList(counts = t(Y), group = X1)
dge <- edgeR::calcNormFactors(dge)
MWUTest(dge, design)
}
\author{
Jakob Walter
}
